agent_template_52318/requirements.txt
```
python-dotenv
os
langchain-anthropic
langgraph
langchain-community
langchain-core
```